requirements_dev.txt
atomicwrites==1.3.0
attrs==19.1.0
bleach==3.1.0
chardet==3.0.4
coverage==4.5.3
docutils==0.14
idna==2.8
more-itertools==7.0.0
pkginfo==1.5.0.1
pluggy==0.11.0
py==1.8.0
Pygments==2.4.2
pytest==4.5.0
pytest-cov==2.7.1
PyYAML==5.1
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
six==1.12.0
tqdm==4.32.1
twine==1.13.0
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
